requirements.txt
# Core dependencies
pydantic>=2.0.0
PyYAML>=6.0

# LLM providers
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Document processing
beautifulsoup4>=4.12.0
markdownify>=0.11.0
requests>=2.31.0
pypdf2>=3.0.0

# Optional: PDF support
# pdfplumber>=0.10.0
